deploy: bfe0000de95636802aa41dfe0f10688ac798d05a

+++ math-P/interpolation/lagrange1.hpp
@@ -4,8 +4,8 @@
 
 // @description 拉格朗日插值
 
-ll lagrange(const V<ll> &x, const V<ll> &y, ll k) {
-	ll sum = 0, n = x.size();
+int lagrange(const V<int> &x, const V<int> &y, int k) {
+	int sum = 0, n = x.size();
 	for (int i = 0; i < n; i++) {
 		int s1 = 1, s2 = 1;
 		for (int j = 0; j < n; j++) {
